Arquivo de teste variavel STATIC

+++ Biblioteca_Header/Matematica.cpp
@@ -23,7 +23,7 @@ int multiplicacao(int num1, int num2)
 
 float divisao(int num1, int num2)
 {
-	return num1 / num2;
+	return (float)num1 / (float)num2;
 
 }
 
@@ -33,11 +33,11 @@ int potencia(int num1, int num2)
 	return pow (num1, num2);
 
 }
-/*
+
 float raizquadrada(int num1)
 {
 	return sqrt (num1);
 
 }
-*/
+
